refactor: GameLog change domain to view
